+++ tests/NotifiableTraitTest.php
@@ -8,6 +8,9 @@ use Orchestra\Support\Facades\Notifier;
 
 class NotifiableTraitTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * Setup the test environment.
+     */
     public function setUp()
     {
         $app = new Container;
@@ -15,18 +18,28 @@ class NotifiableTraitTest extends \PHPUnit_Framework_TestCase
         Facade::clearResolvedInstances();
         Facade::setFacadeApplication($app);
     }
+
+    /**
+     * Teardown the test environment.
+     */
     public function tearDown()
     {
         m::close();
     }
 
+    /**
+     * Test \Orchestra\Notifier\NotifiableTrait::sendNotification()
+     * method.
+     *
+     * @test
+     */
     public function testSendNotificationTraitMethod()
     {
         $user = m::mock('\Orchestra\Notifier\RecipientInterface', '\Illuminate\Support\Contracts\ArrayableInterface');
         $notifier = m::mock('\Orchestra\Notifier\NotifierInterface');
         $stub = new Notifiable;
 
-        $user->shouldReceive('toArray')->once()->andReturn([
+        $user->shouldReceive('toArray')->twice()->andReturn([
             'id' => 2,
             'fullname' => 'Administrator',
         ]);